+++ protected/humhub/modules/file/widgets/FilePreview.php
@@ -18,6 +18,9 @@ class FilePreview extends \humhub\widgets\JsWidget
     public $hideImageFileInfo = false;
     public $edit = false;
     public $visible = false;
+    
+    public $preventPopover = false;
+    public $popoverPosition = 'right';
 
     /**
      * Draws the Upload Button output.
@@ -37,6 +40,8 @@ class FilePreview extends \humhub\widgets\JsWidget
     public function getData()
     {
         return [
+            'prevent-popover' => $this->preventPopover,
+            'popover-position' => $this->popoverPosition,
             'file-edit' => $this->edit,
             'hide-image-file-info' => $this->hideImageFileInfo
         ];
@@ -45,10 +50,13 @@ class FilePreview extends \humhub\widgets\JsWidget
     protected function getFileData()
     {
         $files = $this->getFiles();
+        
         $result = [];
 
         foreach ($files as $file) {
-            $result[] = FileHelper::getFileInfos($file);
+            if($file) {
+                $result[] = FileHelper::getFileInfos($file);
+            }
         }
 
         return $result;